casts and format specifiers in debug output
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V5-6-patches@19578 06827809-a52a-0410-b366-d66718629ded

+++ agent/mibgroup/host/data_access/swinst_pkginfo.c
@@ -133,7 +133,7 @@ netsnmp_swinst_arch_load( netsnmp_container *container, u_int flags)
     closedir( d );
 
     DEBUGMSGTL(("swinst:load:arch"," loaded %d entries\n",
-                CONTAINER_SIZE(container)));
+                (int)CONTAINER_SIZE(container)));
 
     return 0;
 }
